Initial commit
First commit showing the csr files needed to generate certificates for a
simple internal CA with a signed server and client certificates.

The README file has the commands. Should put those into GitHub actions
at some point to verify that it still works.

Was getting OpenSSL verification error for ages:

```
openssl error 18 at 0 depth lookup: self-signed certificate
```

Trick was figuring out that OpenSSL needs the CN to be different for the
CA and any certificates that it's validating and that cfssl needs the CN
set at the top level of the JSON file rather than in the "names"
section.

Issue(s): None